:zap: gatsby-nodeのTS化
