+++ package.json
@@ -34,6 +34,7 @@
   },
   "devDependencies": {
     "prettier": "2.4.1",
+    "ts-node": "^10.2.1",
     "typescript": "^4.4.3"
   },
   "homepage": "https://github.com/gatsbyjs/gatsby-starter-blog#readme",
